Memory utils finished, and preparing for fully auto gameplay by testing out ideas

+++ vec2.h
@@ -21,8 +21,8 @@ public:
 	}
 	~vec2() {}
 
-	double magnitude() {
-		return std::sqrt(x * x + y * y);
+	unsigned magnitude() {
+		return (unsigned)(std::sqrt(x * x + y * y));
 	}
